Video_titles_generator/requirements.txt
ggoogle-generativeai>=0.3.0
pathlib>=1.0.1
python-dotenv
